+++ lib/langchain/tool/calculator.rb
@@ -2,11 +2,10 @@
 
 module Langchain::Tool
   #
-  # A calculator tool that falls back to the Google calculator widget
+  # A calculator tool
   #
   # Gem requirements:
   #     gem "eqn", "~> 1.6.5"
-  #     gem "google_search_results", "~> 2.0.0"
   #
   # Usage:
   #     calculator = Langchain::Tool::Calculator.new
@@ -15,7 +14,7 @@ module Langchain::Tool
     extend Langchain::ToolDefinition
     include Langchain::DependencyHelper
 
-    define_function :execute, description: "Evaluates a pure math expression or if equation contains non-math characters (e.g.: \"12F in Celsius\") then it uses the google search calculator to evaluate the expression" do
+    define_function :execute, description: "Evaluates a pure math expression" do
       property :input, type: "string", description: "Math expression", required: true
     end
 
@@ -23,16 +22,17 @@ module Langchain::Tool
       depends_on "eqn"
     end
 
-    # Evaluates a pure math expression or if equation contains non-math characters (e.g.: "12F in Celsius") then it uses the google search calculator to evaluate the expression
+    # Evaluates a pure math expression
     #
     # @param input [String] math expression
-    # @return [String] Answer
+    # @return [Langchain::Tool::Response] Answer
     def execute(input:)
       Langchain.logger.debug("#{self.class} - Executing \"#{input}\"")
 
-      Eqn::Calculator.calc(input)
+      result = Eqn::Calculator.calc(input)
+      tool_response(content: result)
     rescue Eqn::ParseError, Eqn::NoVariableValueError
-      "\"#{input}\" is an invalid mathematical expression"
+      tool_response(content: "\"#{input}\" is an invalid mathematical expression")
     end
   end
 end