+++ spec/langchain/tool/file_system_spec.rb
@@ -10,13 +10,15 @@ RSpec.describe Langchain::Tool::FileSystem do
     it "lists a directory" do
       allow(Dir).to receive(:entries).with(directory_path).and_return(entries)
       response = subject.list_directory(directory_path: directory_path)
-      expect(response).to eq(entries)
+      expect(response).to be_a(Langchain::ToolResponse)
+      expect(response.content).to eq(entries)
     end
 
     it "returns a no such directory error" do
       allow(Dir).to receive(:entries).with(directory_path).and_raise(Errno::ENOENT)
       response = subject.list_directory(directory_path: directory_path)
-      expect(response).to eq("No such directory: #{directory_path}")
+      expect(response).to be_a(Langchain::ToolResponse)
+      expect(response.content).to eq("No such directory: #{directory_path}")
     end
   end
 
@@ -28,13 +30,15 @@ RSpec.describe Langchain::Tool::FileSystem do
       it "successfully writes" do
         allow(File).to receive(:write).with(file_path, content)
         response = subject.write_to_file(file_path: file_path, content: content)
-        expect(response).to eq(nil)
+        expect(response).to be_a(Langchain::ToolResponse)
+        expect(response.content).to eq("File written successfully")
       end
 
       it "returns a permission denied error" do
         allow(File).to receive(:write).with(file_path, content).and_raise(Errno::EACCES)
         response = subject.write_to_file(file_path: file_path, content: content)
-        expect(response).to eq("Permission denied: #{file_path}")
+        expect(response).to be_a(Langchain::ToolResponse)
+        expect(response.content).to eq("Permission denied: #{file_path}")
       end
     end
 
@@ -45,13 +49,15 @@ RSpec.describe Langchain::Tool::FileSystem do
       it "successfully reads" do
         allow(File).to receive(:read).with(file_path).and_return(content)
         response = subject.read_file(file_path: file_path)
-        expect(response).to eq(content)
+        expect(response).to be_a(Langchain::ToolResponse)
+        expect(response.content).to eq(content)
       end
 
       it "returns an error" do
         allow(File).to receive(:read).with(file_path).and_raise(Errno::ENOENT)
         response = subject.read_file(file_path: file_path)
-        expect(response).to eq("No such file: #{file_path}")
+        expect(response).to be_a(Langchain::ToolResponse)
+        expect(response.content).to eq("No such file: #{file_path}")
       end
     end
   end